Explicitly cast from int to double

+++ src/corecel/grid/UniformGrid.hh
@@ -80,7 +80,7 @@ UniformGrid::UniformGrid(UniformGridData const& data) : data_(data)
 CELER_FUNCTION auto UniformGrid::operator[](size_type i) const -> value_type
 {
     CELER_EXPECT(i < data_.size);
-    return data_.front + data_.delta * i;
+    return data_.front + data_.delta * static_cast<double>(i);
 }
 
 //---------------------------------------------------------------------------//
